# lib/api-zod/src/generated/types/commsChannel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synvet — Plataforma SaaS Veterinária
 * OpenAPI spec version: 0.1.0
 */
import type { CommsChannelKind } from "./commsChannelKind";
import type { CommsChannelMeta } from "./commsChannelMeta";
import type { CommsChannelStatus } from "./commsChannelStatus";

export interface CommsChannel {
  id: string;
  clinicId: string;
  kind: CommsChannelKind;
  provider: string;
  status: CommsChannelStatus;
  phoneNumber?: string | null;
  displayName?: string | null;
  externalId?: string | null;
  lastConnectedAt?: Date | null;
  lastError?: string | null;
  meta: CommsChannelMeta;
  createdAt: Date;
  updatedAt: Date;
}