# lib/api-zod/src/generated/types/commsTestSendBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synvet — Plataforma SaaS Veterinária
 * OpenAPI spec version: 0.1.0
 */

export interface CommsTestSendBody {
  channelId: string;
  /** @minLength 5 */
  toAddress: string;
  /**
   * @minLength 1
   * @maxLength 4000
   */
  body: string;
}