# lib/api-zod/src/generated/types/createCommsAutomationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synvet — Plataforma SaaS Veterinária
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCommsAutomationBodyConfig } from "./createCommsAutomationBodyConfig";

export interface CreateCommsAutomationBody {
  /**
   * @minLength 2
   * @maxLength 120
   */
  name: string;
  trigger: string;
  templateId: string;
  channelId?: string | null;
  offsetMinutes?: number;
  config?: CreateCommsAutomationBodyConfig;
  enabled?: boolean;
}