# lib/api-zod/src/generated/types/listCommsMessagesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synvet — Plataforma SaaS Veterinária
 * OpenAPI spec version: 0.1.0
 */
import type { CommsMessage } from "./commsMessage";

export interface ListCommsMessagesResponse {
  items: CommsMessage[];
}