# lib/api-zod/src/generated/types/updateCommsTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Synvet — Plataforma SaaS Veterinária
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCommsTemplateBody {
  name?: string;
  channel?: string;
  category?: string;
  body?: string;
  variables?: string[];
  enabled?: boolean;
}